+++ lib/router/routes.dart
@@ -7,6 +7,7 @@ import '../screens/games/game/game_screen.dart';
 import '../screens/games/games_screen.dart';
 import '../screens/games/new/new_game_screen.dart';
 import '../screens/login/login.dart';
+import '../screens/my_account/my_account_screen.dart';
 import '../screens/set_up_profile/set_up_profile_screen.dart';
 import '../screens/settings/settings_screen.dart';
 import '../screens/sign_up/sign_up.dart';
@@ -31,7 +32,7 @@ class HomeRoute extends GoRouteData {
 }
 
 @TypedGoRoute<LoginRoute>(
-  path: '/login',
+  path: '/${LoginRoute.path}',
   routes: [
     TypedGoRoute<SignUpRoute>(path: 'sign-up'),
   ],
@@ -42,6 +43,8 @@ class LoginRoute extends GoRouteData {
   });
 
   final Uri? redirectUrl;
+
+  static const path = 'login';
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
@@ -58,9 +61,11 @@ class SignUpRoute extends GoRouteData {
   }
 }
 
-@TypedGoRoute<VerifyEmailRoute>(path: '/verify-email')
+@TypedGoRoute<VerifyEmailRoute>(path: '/${VerifyEmailRoute.path}')
 class VerifyEmailRoute extends GoRouteData {
   const VerifyEmailRoute();
+
+  static const path = 'verify-email';
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
@@ -68,19 +73,11 @@ class VerifyEmailRoute extends GoRouteData {
   }
 }
 
-@TypedGoRoute<SetUpProfileRoute>(path: '/set-up-profile')
-class SetUpProfileRoute extends GoRouteData {
-  const SetUpProfileRoute();
-
-  @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const SetUpProfileScreen();
-  }
-}
-
-@TypedGoRoute<WaitingForApprovalRoute>(path: '/waiting-for-approval')
+@TypedGoRoute<WaitingForApprovalRoute>(path: '/${WaitingForApprovalRoute.path}')
 class WaitingForApprovalRoute extends GoRouteData {
   const WaitingForApprovalRoute();
+
+  static const path = 'waiting-for-approval';
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
@@ -88,18 +85,66 @@ class WaitingForApprovalRoute extends GoRouteData {
   }
 }
 
-@TypedGoRoute<SettingsRoute>(path: '/settings')
-class SettingsRoute extends GoRouteData {
-  const SettingsRoute();
+@TypedGoRoute<SetUpProfileRoute>(path: '/${SetUpProfileRoute.path}')
+class SetUpProfileRoute extends GoRouteData {
+  const SetUpProfileRoute();
+
+  static const path = 'set-up-profile';
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
-    return const SettingsScreen();
+    return const SetUProfileScreen();
+  }
+}
+
+@TypedGoRoute<SettingsRoute>(
+  path: '/${SettingsRoute.path}',
+  routes: [
+    TypedGoRoute<MyAccountRoute>(
+      path: MyAccountRoute.path,
+    )
+  ],
+)
+class SettingsRoute extends GoRouteData {
+  const SettingsRoute();
+
+  static const path = 'settings';
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const SettingsScreen(
+      public: false,
+    );
+  }
+}
+
+@TypedGoRoute<PublicSettingsRoute>(
+  path: '/${PublicSettingsRoute.path}',
+)
+class PublicSettingsRoute extends GoRouteData {
+  const PublicSettingsRoute();
+
+  static const path = 'public-settings';
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const SettingsScreen(public: true);
+  }
+}
+
+class MyAccountRoute extends GoRouteData {
+  const MyAccountRoute();
+
+  static const path = 'my-account';
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const MyAccountScreen();
   }
 }
 
 @TypedGoRoute<GamesRoute>(
-  path: '/games',
+  path: '/${GamesRoute.path}',
   routes: [
     TypedGoRoute<NewGameRoute>(path: 'new'),
     TypedGoRoute<GameRoute>(path: ':id'),
@@ -107,6 +152,8 @@ class SettingsRoute extends GoRouteData {
 )
 class GamesRoute extends GoRouteData {
   const GamesRoute();
+
+  static const path = 'games';
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
@@ -142,9 +189,11 @@ class GameRoute extends GoRouteData {
   }
 }
 
-@TypedGoRoute<UsersRoute>(path: '/users')
+@TypedGoRoute<UsersRoute>(path: '/${UsersRoute.path}')
 class UsersRoute extends GoRouteData {
   const UsersRoute();
+
+  static const path = 'users';
 
   @override
   Future<String?> redirect(BuildContext context, GoRouterState state) async {