+++ lib/screens/settings/settings_button.dart
@@ -2,15 +2,15 @@ import 'package:flutter/material.dart';
 
 import '../../router/routes.dart';
 
-class SettingsButton extends StatelessWidget {
-  const SettingsButton({super.key});
+class PublicSettingsButton extends StatelessWidget {
+  const PublicSettingsButton({super.key});
 
   @override
   Widget build(BuildContext context) {
     return IconButton(
       icon: const Icon(Icons.settings),
       onPressed: () {
-        const SettingsRoute().push(context);
+        const PublicSettingsRoute().push(context);
       },
     );
   }