+++ lib/screens/settings/settings_screen.dart
@@ -7,17 +7,24 @@ import '../../services/auth.dart';
 import '../../services/auth_user_notifier.dart';
 import '../../services/info_plus.dart';
 import '../../utils/bool.dart';
+import '../../widgets/beta_widget.dart';
+import '../../widgets/bottom_navigation.dart';
 import '../../widgets/theme_switch.dart';
 
 class SettingsScreen extends ConsumerWidget {
-  const SettingsScreen({super.key});
+  const SettingsScreen({
+    required this.public,
+    super.key,
+  });
+
+  final bool public;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final isLoggedIn = ref.watch(loggedInAuthUserProvider).isLoggedIn;
     return Scaffold(
       appBar: AppBar(
-        leading: (!Navigator.of(context).canPop()).nullIfFalse(
+        leading: (public && !Navigator.of(context).canPop()).nullIfFalse(
           IconButton(
             onPressed: () {
               const HomeRoute().go(context);
@@ -29,17 +36,22 @@ class SettingsScreen extends ConsumerWidget {
       ),
       body: ListView(
         children: [
-          const ListTile(
-            title: Row(
-              children: [
-                Expanded(
-                  child: Text('Theme'),
-                ),
-                ThemeSwitch()
-              ],
+          if (!public && isLoggedIn)
+            BetaWidget(
+              child: ListTile(
+                leading: const Icon(Icons.person),
+                onTap: () => const MyAccountRoute().go(context),
+                title: const Text('My account'),
+                trailing: const Icon(Icons.chevron_right),
+              ),
             ),
+          const ListTile(
+            leading: Icon(Icons.contrast),
+            title: Text('Theme'),
+            trailing: ThemeSwitch(),
           ),
           AboutListTile(
+            icon: const Icon(Icons.info),
             applicationVersion:
                 ref.watch(infoPlusProvider).valueOrNull?.version ?? '',
           ),
@@ -61,6 +73,7 @@ class SettingsScreen extends ConsumerWidget {
           ],
         ],
       ),
+      bottomNavigationBar: (!public).nullIfFalse(const BottomNavigation()),
     );
   }
 }