+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adds an about page.
 - Add licenses.
 - Remove the about page and includes everything in the settings page.
+- Adds a bottom navigation bar.
 
 ## 0.16.0 [2023-11-02]